# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "2Gather API DOC",
		"description": "OpenAPI Specification for 2Gather API"
	},
	"schemes": [
		"http"
	],
	"host": "localhost:4003",
	"basePath": "/api",
...
	"paths": {
...
		"/device": {
			"put": {
				"tags": [
					"Device"
				],
				"summary": "Set user device token",
				"description": "Functionality for users to set device token.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"in": "body",
						"name": "Set device token",
						"description": "Set device token request body",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"deviceToken": {
									"type": "string"
								},
								"deviceId": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/device/language": {
			"put": {
				"tags": [
					"Device"
				],
				"summary": "Set user device language",
				"description": "Functionality for users to set device language.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/languageInHeader"
					},
					{
						"in": "body",
						"name": "Set device language",
						"description": "Set device language request body",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"deviceId": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
...
		"/staff": {
			"post": {
				"tags": [
					"Staff"
				],
				"summary": "Create new staff member",
				"description": "Create new staff member functionality for admin",
				"consumes": "application/x-www-form-urlencoded",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"name": "translations[0]",
						"description": "Translation list(JSON Stringified { name: string, language: number, occupation: string, description: string })",
						"type": "string",
						"in": "formData",
						"required": true
					},
					{
						"name": "facebookAccount",
						"description": "Facebook uri",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "linkedInAccount",
						"description": "LinkedIn uri",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "twitterAccount",
						"description": "Twitter uri",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "avatar",
						"description": "Avatar of the staff member",
						"type": "file",
						"in": "formData",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"put": {
				"tags": [
					"Staff"
				],
				"summary": "Update staff member",
				"description": "Update staff member functionality for admin",
				"consumes": "application/x-www-form-urlencoded",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"name": "id",
						"description": "Staff member id",
						"type": "string",
						"in": "formData",
						"required": true
					},
					{
						"name": "translations[0]",
						"description": "Translation list(JSON Stringified { name: string, language: number, occupation: string, description: string })",
						"type": "string",
						"in": "formData",
						"required": true
					},
					{
						"name": "facebookAccount",
						"description": "Facebook uri",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "linkedInAccount",
						"description": "LinkedIn uri",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "twitterAccount",
						"description": "Twitter uri",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "avatar",
						"description": "Avatar of the staff member",
						"type": "file",
						"in": "formData",
						"required": false
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"delete": {
				"tags": [
					"Staff"
				],
				"summary": "Delete staff member",
				"description": "Delete staff member functionality for admins.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/idInQuery"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/staff/adminList": {
			"post": {
				"tags": [
					"Staff"
				],
				"summary": "Get staff member list for admin",
				"description": "Functionality to get staff member list for admin.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"in": "body",
						"name": "Get staff member list",
						"description": "Get staff member list body.",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"pageNo": {
									"type": "number"
								},
								"limit": {
									"type": "number"
								},
								"search": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/staff/status": {
			"put": {
				"tags": [
					"Staff"
				],
				"summary": "Update staff member status for admin",
				"description": "Functionality to update staff members status for admin.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"in": "body",
						"name": "Change staff member status",
						"description": "Change staff member status body.",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"id": {
									"type": "string"
								},
								"status": {
									"type": "number"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/staff/adminDetails": {
			"get": {
				"tags": [
					"Staff"
				],
				"summary": "Get staff member details(ADMIN)",
				"description": "Get staff member functionality for admins.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/idInQuery"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/staff/list": {
			"get": {
				"tags": [
					"Staff"
				],
				"summary": "Get staff member list(USER)",
				"description": "Get staff member list for users.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/languageInHeader"
					},
					{
						"$ref": "#/parameters/pageNo"
					},
					{
						"$ref": "#/parameters/limit"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/partner": {
			"post": {
				"tags": [
					"Partner"
				],
				"summary": "Create new partner",
				"description": "Create new partner functionality for admin",
				"consumes": "application/x-www-form-urlencoded",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"name": "translations[0]",
						"description": "Translation list(JSON Stringified { name: string, language: number })",
						"type": "string",
						"in": "formData",
						"required": true
					},
					{
						"name": "contactPerson",
						"description": "Contact Person",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "phone",
						"description": "Phone number",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "avatar",
						"description": "Avatar of the partner",
						"type": "file",
						"in": "formData",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"put": {
				"tags": [
					"Partner"
				],
				"summary": "Update partner",
				"description": "Update partner functionality for admin",
				"consumes": "application/x-www-form-urlencoded",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"name": "id",
						"description": "Partner Id",
						"type": "string",
						"in": "formData",
						"required": true
					},
					{
						"name": "translations[0]",
						"description": "Translation list(JSON Stringified { name: string, language: number })",
						"type": "string",
						"in": "formData",
						"required": true
					},
					{
						"name": "contactPerson",
						"description": "Contact Person",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "phone",
						"description": "Phone Number",
						"type": "string",
						"in": "formData",
						"required": false
					},
					{
						"name": "avatar",
						"description": "Avatar of the partner",
						"type": "file",
						"in": "formData",
						"required": false
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"get": {
				"tags": [
					"Partner"
				],
				"summary": "Get partner details",
				"description": "Get partner details functionality for admins.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/idInQuery"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"delete": {
				"tags": [
					"Partner"
				],
				"summary": "Delete partner",
				"description": "Delete partner functionality for admins.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/idInQuery"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/partner/adminList": {
			"post": {
				"tags": [
					"Partner"
				],
				"summary": "Get partner list for admin",
				"description": "Functionality to get partner list for admin.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"in": "body",
						"name": "Get partner list",
						"description": "Get partner list body.",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"pageNo": {
									"type": "number"
								},
								"limit": {
									"type": "number"
								},
								"search": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
...
	},
	"parameters": {
		"Authorization": {
			"name": "Authorization",
			"in": "header",
			"required": true,
			"description": "Bearer token",
			"type": "string"
		},
...
		"idInQuery": {
			"name": "id",
			"in": "query",
			"required": true,
			"description": "Item Id",
			"type": "string"
		},
...
		"pageNo": {
			"name": "pageNo",
			"in": "query",
			"required": true,
			"description": "Page number",
			"type": "number"
		},
		"limit": {
			"name": "limit",
			"in": "query",
			"required": true,
			"description": "Limit to get",
			"type": "number"
		},
...
		"languageInHeader": {
			"name": "language",
			"in": "header",
			"required": true,
			"description": "Language",
			"type": "number"
		}
	}
}